// Sorgenti GARI relativi al rilascio del 25-06-2025/GARI_Umbria_2025-06-25/GiasNG/src/app/Model/attivita/centri_di_costo/CapoAnimaleCDC.ts
import { BaseCodeDescr, CapoAnimale, SottogruppoStalla } from "app/Service/api.service";
import { CentroDiCosto, Tipo } from "./CentroDiCosto";

export class CapoAnimaleCDC extends CentroDiCosto {

    sottogruppoStalla_ingresso: SottogruppoStalla;

    sottogruppoStalla_uscita: SottogruppoStalla;

    capoAnimale: CapoAnimale;

    capoAnimaleNonPresente: BaseCodeDescr;

    id_movimentazione_BDN: number;

    qtaSomministrata: number;

    constructor() {
        super();
        this.classType = "CapoAnimaleCDC";
        this.tipo = Tipo.CapoAnimale;
    }

}