// Sorgenti GARI relativi al rilascio del 25-06-2025/GARI_Umbria_2025-06-25/GiasNG/src/app/Model/TipiEnumerativi.ts
export enum enum_PagineGiasNG {
  Pagina_Corrente = 0,
  Pagina_Menu_Anagrafica = 1,

  Pagina_Menu_Anagrafica_Imprese = 10,
  Pagina_Menu_Anagrafica_Centri = 11,
  Pagina_Menu_Anagrafica_Catasto = 12,
  Pagina_Menu_Anagrafica_Campi = 13,
  Pagina_Menu_Anagrafica_Impianti = 14,
  Pagina_Menu_Anagrafica_Macchine = 15,
  Pagina_Menu_Anagrafica_Contatti = 16,
  Pagina_Menu_Anagrafica_Fabbricati = 17,

  Pagina_Edit_Impresa = 2,
  Pagina_Edit_Centro = 3,
  Pagina_Edit_Campo = 4,
  Pagina_Edit_Catasto = 5,
  Pagina_Edit_AppezzamentoGlobal = 6,
  Pagina_Edit_Fabbricato = 7,
  Pagina_Edit_Contatto = 8,
  Pagina_Edit_Macchina = 9,

  Pagina_Menu_Agenda = 20,
  Pagina_Edit_Attivita = 21,
  Pagina_Configurazione_Operazioni_Culturali = 22,

  Pagina_Menu_Profilazione = 30,
  Pagina_Profilazione_Import_Utenti = 31,

  Pagina_Menu_Gruppi_Merce_Anagrafica = 40,
  Pagina_Menu_Gruppi_Merce_Autorizzazioni = 41,

  Pagina_AmministrazioneSistema_ConsultaSincroDatiApp = 50,

  Pagina_Menu_Visite = 60,
  Pagina_Edit_Visite = 61,

  Pagina_Valutazioni = 70,
  Pagina_Valutazioni_Edit = 71,

  Pagina_PianoConti = 80,

  Pagina_Menu_Zoo = 90,

  Pagina_Budget_Menu_Anagrafica_Imprese = 100,
  Pagina_Budget_Menu_Anagrafica_Centri = 110,
  Pagina_Budget_Menu_Anagrafica_Catasto = 120,
  Pagina_Budget_Menu_Anagrafica_Campi = 130,
  Pagina_Budget_Menu_Anagrafica_Impianti = 140,
  Pagina_Budget_Menu_Anagrafica_Macchine = 150,

  Pagina_GIS = 200,
  Pagina_GIS_cfg_proiezioni = 201,

  Pagina_Dashboard = 999,
  Pagina_Preferiti_Config = 1000,

  Pagina_Gruppi_Raccolta = 300,
  Pagina_Requisiti_Stabilimento = 301,
  Pagina_Requisiti_Stabilimento_PianoColturale = 302,
  Pagina_Requisiti_Stabilimento_Contratti = 303,
  Pagina_Dati_Previsionali_Colture = 305,
  Pagina_Requisiti_Stabilimento_VD_PianoColturale = 306,
  Pagina_Requisiti_Stabilimento_VD_Contratti = 307,
  Pagina_Confronto_Piano_Colturale = 308,
  Pagina_Filtro_Ricerca = 309,
  Pagina_Export_QdC_To_Agea = 310,
  Pagina_Profilazione_Imprese = 311,
  Pagina_Analisi_Terreno_Menu = 312,
  Pagina_Analisi_Terreno_Edit = 313,
  Pagina_Report_Impiego_Prodotti_Fitosanitari = 314,
  Pagina_Lettura_Contatori = 315,

  Pagina_Domanda_Irrigua = 320,
  Pagina_Terapie = 321,
  Pagina_Menu_Rilievi = 322,

  Pagina_Trattamento_Zoo = 323
}